backend/pyproject.toml
[project]
name = "ragchatbot"
version = "0.1.0"
description = "A RAG-based chatbot for book content using Gemini API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.23.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.40.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "qdrant-client>=1.9.0",
    "google-generativeai>=0.5.0",
    "cohere>=5.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"